🚚 Rename: atom -> myprofileTechState로 변경
